+++ changelog.md
@@ -1,3 +1,5 @@
 
 # Changes From Design Document
 - Removed the `gold` attribute from the `Fight` class.
+- Added `getScore` abstract method to the `User` class.
+- Removed `checkRanking` method from the `User` class.